bitcoin: move bip158 test to tests/
Move blockfilters to tests/ because it include_str's
data from tests/, which we exclude when publishing.
